requirements.txt
numpy==1.23.2
pandas==1.4.3
boto3==1.24.24
requests==2.25.1
awswrangler==3.3.0
pyarrow==13.0.0
openpyxl==3.1.2
python-dotenv==0.10.1